# logplex/encoding: Improve encoding performance