--- a/logplex/encoding/encoder.go
+++ b/logplex/encoding/encoder.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"io"
 	"strconv"
+	"strings"
 
 	"github.com/pkg/errors"
 )
@@ -91,7 +92,16 @@ func (s *sseEncoder) separator() {
 }
 
 func messageToString(msg Message) string {
-	return fmt.Sprintf("%s %s[%s]: %s", msg.Timestamp.Format(HumanTimeFormat), msg.Application, msg.Process, msg.Message)
+	// equivalent to fmt.Sprintf("%s %s[%s]: %s", msg.Timestamp.Format(HumanTimeFormat), msg.Application, msg.Process, msg.Message)
+	var s strings.Builder
+	s.WriteString(msg.Timestamp.Format(HumanTimeFormat))
+	s.WriteString(" ")
+	s.WriteString(msg.Application)
+	s.WriteString("[")
+	s.WriteString(msg.Process)
+	s.WriteString("]: ")
+	s.WriteString(msg.Message)
+	return s.String()
 }
 
 // Encode serializes a syslog message into their wire format ( octet-framed syslog )