@freeformz was so kind to point out room for improvement here. Use `strings.Builder` instead of `fmt.Sprintf` to format logplex messages.