@cyx 2020-05-08
Nice!

On Thu, May 7, 2020, 15:29 ypaq <notifications@github.com> wrote:

> @freeformz <https://github.com/freeformz> was so kind to point out room
> for improvement here. Use strings.Builder instead of fmt.Sprintf to
> format logplex messages.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/heroku/x/pull/135
> Commit Summary
>
>    - Improve messageToString performance
>
> File Changes
>
>    - *M* logplex/encoding/encoder.go
>    <https://github.com/heroku/x/pull/135/files#diff-f571490d3e1559781ad1c0bfb00e7e8c>
>    (12)
>    - *M* logplex/encoding/encoder_test.go
>    <https://github.com/heroku/x/pull/135/files#diff-d40c6fc38d04fb823c35cdb9ea06c875>
>    (19)
>
> Patch Links:
>
>    - https://github.com/heroku/x/pull/135.patch
>    - https://github.com/heroku/x/pull/135.diff
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/heroku/x/pull/135>, or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAAARUP2GFPYJZOARF7AO6LRQMY4LANCNFSM4M3WWHFA>
> .
>