@cyx 2020-05-08
#winning

On Thu, May 7, 2020, 20:34 ypaq <notifications@github.com> wrote:

> I have not, thank you!
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/heroku/x/pull/135#issuecomment-625611682>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAAARUOPDY72CBFRMQQ77NDRQN4T5ANCNFSM4M3WWHFA>
> .
>